Cleanup script output
The "Done" message was often not being output because
of the 'continue' statements earlier in the loop.
The "Done" is pointless anyway - it's obvious from the
message that will follow immediately after (next batch,
or complete) that the batch is done anyway.

Also moved waitforreplication - which was also being
skipped because of the 'continue's earlier on - up top.

Change-Id: I6ec3bfa0e269de07b32e900d46d7f5adeba4c7db