Use getOption to detect user preferences
The problem with doing it right in the query is that
that database may not have all information.
GlobalPreferences, if installed, may live in another
database. Using the built-in `getOption` is safer,
since GlobalPreferences, if installed, will hook into
it and add global preferences in there as well.

This makes things a little less efficient for us,
because we'll now have to query 2 databases (or caches)
for every user, instead of being able to do it all in
1 go. That said, we're only looking at a few thousand
users realistically, so it might not even have too much
of an impact. And the query has also become simpler,
with fewer joins and conditions.

Bug: T313209
Change-Id: Ic4dd446f2eed3b1716655143b92bf1d3a2d0652d