Address performance issue in ImageSuggestions
The ImageSuggestions module is incorrectly loading modules
to all article page views. When CSS is unused

I am not aware of how this feature works but it seems like
it may be possible to at least
 limit the module to logged in users.

On long term this should be further scoped to when the icons
are needed.

Bug: T403771
Change-Id: I91e40cadfa9fd5e796e29e1e08a1fa1cb728cbca